tsにした(動くかは知らん！
