+++ functions/.gitignore
@@ -1,0 +1,4 @@
+node_modules/
+
+src/**/*.js
+src/**/*.js.map
