+++ functions/package.json
@@ -1,10 +1,23 @@
 {
   "name": "silvershark",
   "description": "Presence for Firestore",
+  "private": true,
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "build": "tsfmt -r && tsc",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
     "@google-cloud/pubsub": "^0.16.2",
     "firebase-admin": "^5.0.0",
     "firebase-functions": "^0.7.0"
   },
-  "private": true
+  "devDependencies": {
+    "@types/google-cloud__pubsub": "^0.14.0",
+    "typescript": "^2.7.1",
+    "typescript-formatter": "^7.0.1"
+  }
 }
